Don't allow pipeline input.
The reason for this decision is because running LogParser multiple times
for each user/device in the pipeline would be pretty inefficient.  There
are clearly ways to make pipelining work (run LogParser and return all
results in BEGIN, then parse through them in the PROCESS block.  But I
haven't written that yet.